+++ client/src/Components/Footer.jsx
@@ -3,6 +3,7 @@ import { Link } from 'react-router-dom';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faLinkedin, faInstagram, faFacebook } from '@fortawesome/free-brands-svg-icons';
 
+
 const Footer = () => {
   const currentYear = new Date().getFullYear();
